Set telinit path to /lib/sysvinit/telinit
If systemd is not the active PID 1 and the SysV compat tools are
provided by systemd-sysv, /sbin/telinit simply forwards the requests to
the native telinit implementation.
To make the upgrade from sysvinit to systemd less disruptive, the plan
is to make the transitional sysvinit package provide
/lib/sysvinit/telinit and /lib/sysvinit/init which then can be used as a
fallback in case of emergency.
Set the telinit path in systemd accordingly in preparation for that.