Properly escape network interface names in net.agent
Use the new systemd-escape utility to properly escape the network
interface name when starting an ifup@.service instance for hotplugged
network interfaces. Make sure a recent enough systemd version is
installed by bumping the versioned Breaks accordingly.

Also make "systemctl start" non-blocking as we only want to enqueue the
job but do not actually need to wait for it to finish.

Closes: #747044