Add userspace firmware loader stub
Reinstate a debian/extra/rules/50-firmware.rules which immediately tells the
kernel that userspace firmware loading failed. Otherwise it tries for a minute
to call the userspace helper (if CONFIG_FW_LOADER_USER_HELPER is enabled) in
vain, which causes long delays with devices which have a range of possible
firmware versions. (LP: #1398458)