Use common-session-noninteractive in systemd-user's PAM config
Instead of common-session. The latter can include PAM modules like libpam-mount
which expect to be called just once and/or interactively, which already happens
for login, ssh, or the display-manager. Add pam_systemd.so explicitly, as it's
not included in -noninteractive, but is always required (and idempotent).

There is no net change on systemd which don't use manually installed PAM
modules.

Closes: #739676