Revert networkd calling if-{up,post-down}.d/ scripts
About half of the existing hooks are not relevant or even actively detrimental
when running with networkd. For the relevant ones, a lot of them should be
fixed in the projects themselves (using IP_FREEBIND etc.).

This has also caused some random hangs when restarting networkd or on shutdown,
by triggering #774153. While this needs to be fixed anyway, it makes this bug
much more apparent.

Closes: #798625