/lib/lsb/init-functions.d/40-systemd: don't reload masked services
When a service is masked the native service does not support
reload, so the lsb hooks would then allow fallback on init script
reload function.

Move the check for masked services earlier and bail out with exit 0
so that does not happen.

(Still a bit confusing that there's no output at all to inform
the user calling the init script for a masked service about this,
but I've been told it's on purpose eg. to avoid spamming the logs.)

Closes: #804882