Replace "sysv-rc" dependency with Conflicts: openrc, file-rc
The rationale from #739679 still applies, but with the moving of
{invoke,update}-rc.d to init-system-helpers we don't actually need anything
from sysv-rc any more other than the assumption that SysV init scripts are
enabled in /etc/rc?.d/ for the SysV generator to work (and file-rc and openrc
don't do that).

This drops the last remaining dependency on a sysvinit binary.