Add debian/extra/checkout-upstream
This is a script to replace the current source with a checkout of an upstream
pull request, branch, or commit, and remove debian/patches/. This also disables
autopkgtests which are not appropriate for testing upstream. Called from
debian/rules if $TEST_UPSTREAM is set.

This will be used for upstream CI.